chore(views): WIP TS migration
