+++ src/mixins/data-workflows/item-mixin.ts
@@ -1,0 +1,51 @@
+/**
+ * To use this mixin, parent component has to define `moduleName` in data.
+ * It will fetch module name information depending url `id` parameter.
+ */
+
+import { Component, Vue } from 'vue-property-decorator';
+import { State } from 'vuex-class';
+import DataManagementHeader from '@/components/data-workflows/DataManagementHeader.vue';
+import ItemComponent from '@/components/data-workflows/common/item/ItemComponent.vue';
+
+import store from '@/store';
+
+@Component({
+	components: { DataManagementHeader, ItemComponent }
+})
+export default class ItemMixin extends Vue {
+	isNotFound: boolean = false;
+	item: Object = {}; // TODO: Create type `Item`
+	isLoading: boolean = true;
+
+	@State(state => state[this.moduleName].data) firestoreItem: Object = {}; // TODO: Create type `Item`
+
+	mounted() {
+		if (!this.moduleName) throw new Error('Parent component has to define `moduleName` in data');
+		this.getItem();
+	}
+
+	async getItem() {
+		if (!this.firestoreItem[this.itemId]) await this.getFirestoreData();
+		this.item = this.firestoreItem[this.itemId];
+
+		if (!this.item) this.isNotFound = true;
+		this.isLoading = false;
+	}
+	async getFirestoreData() {
+		await store.dispatch(`${this.moduleName}/closeDBChannel`, { clearModule: true });
+		await store.dispatch(`${this.moduleName}/fetchById`, this.itemId);
+	}
+
+	/*computed: {
+		...mapState({
+			firestoreItem(state) {
+				return state[this.moduleName].data;
+			}
+		})
+	}*/
+
+	get itemId(): string {
+		return this.$route.params.id;
+	}
+}
